refactor: keybindings and commands
We were handling keybindings in two different places:
- in the useGlobalKeybindings hook
- in the commands registry

This refactor keeps all the keybindings in a single place and invoking a
keybinding results in finding and invoking a commmand-id.

The command-id and how to invoke the command is in the utils/commands.ts file.

The command palette uses both the information to create the list of commands
including the keybindings for showing it to the user.